// src/client/app/features/nttbalance/nttbalance.ts
export interface INttbalance {
  id: string;
  OexercComptaKey?: string;
  oexercCompta ?: string;
  OtableauposteKey?: string;
  otableauposte?: string;
  OreferenceKey?: string;
  oreference ?: string;
  NumCompte: string;
  IntitulCompte: string;
  SoldeDebit?: number;
  SoldeCredit?: number;
}

export class Nttbalance {
}